% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunu_geodata.R
\name{list_all_geodata}
\alias{list_all_geodata}
\title{All kind of data available}
\usage{
list_all_geodata()
}
\value{
Data available in this package
}
\description{
All kind of data available
}
\examples{
list_all_geodata()
}
