% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunu_geodata.R
\name{sunu_geodata}
\alias{sunu_geodata}
\title{Read geospatial data}
\usage{
sunu_geodata(path)
}
\arguments{
\item{path}{location of data available}
}
\value{
a \code{tibble} with geometry feature
}
\description{
Read geospatial data
}
\examples{
geodata_path<-system.file("extdata","Regionsn.gpkg",package="SunuGeoData")
sunu_geodata(geodata_path)

}
